+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * various `data.ThingID` types now also have a corresponding `data.ThingName` for the union of `string` and all known specific names. Runtime `ThingID` union types have the corresponding `data.ThingName` (with `data.` prefix still) inserted as well.
     * `on_event` and `event_handler` distinguish event name strings between CustomEvent and CustomInputEvent
     * various functions which take a param table with subtypes have been split to overloads
+    * various dicts on LuaPrototypes are populated with names, and `mod_data`s use `data_type` as the LuaLS type for `data`
 * Debugger:
   * `tags` in launch config to pass arbirary (small please) data to lua code as `__DebugAdapter.tags`
   * escape `\n` in strings when displayed